test: move service integration tests to separate directory
