+++ src/test/kotlin/se/uulm/snowballr/backend/integration/services/ReadingListIntegrationTest.kt
@@ -1,9 +1,10 @@
-package se.uulm.snowballr.backend.integration
+package se.uulm.snowballr.backend.integration.services
 
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import se.uulm.snowballr.backend.DataBuilder
+import se.uulm.snowballr.backend.integration.IntegrationTest
 import se.uulm.snowballr.backend.model.EntityType
 import se.uulm.snowballr.backend.model.parseUUID
 import kotlin.test.assertFalse
